Add "emits" to circle and rectangle and register listener to actively propagate events
As for vue3, a component can (should?) have an "emits" option containing all its emits.
For some components this was implemented, but for rectangle and circle it was missing.

Further re-used the addListener code from other components to register listeners to
the google maps api events and propagate them to the partent component of a circle/rectangle.